+++ scripts/calculateRevenue/protocols/fonbnk/index.ts
@@ -1,95 +1,17 @@
 import { BlockField, Query, QueryResponse } from '@envio-dev/hypersync-client'
-import { getErc20Contract, getHyperSyncClient } from '../../../utils'
+import { getBlock, getErc20Contract, getHyperSyncClient } from '../../../utils'
 import {
-  FONBNK_API_URL,
-  FONBNK_CLIENT_ID,
   FonbnkNetwork,
   fonbnkNetworkToNetworkId,
   TRANSACTION_VOLUME_USD_PRECISION,
 } from './constants'
-import { fetchWithBackoff } from '../../../protocolFilters/beefy'
-import { generateSignature } from './helpers'
+import { fetchFonbnkAssets, getPayoutWallets } from './helpers'
 import { paginateQuery } from '../../../utils/hypersyncPagination'
 import { NetworkId } from '../../../types'
 import { fetchTokenPrices } from '../utils/tokenPrices'
 import { getTokenPrice } from '../beefy'
-import { Address } from 'viem'
-
-interface FonbnkAsset {
-  network: FonbnkNetwork
-  asset: string
-}
-
-interface FonbnkTransaction {
-  amount: bigint
-  tokenAddress: Address
-  timestamp: Date
-}
-
-async function fetchFonbnkAssets(): Promise<FonbnkAsset[]> {
-  const url = `${FONBNK_API_URL}/api/pay-widget-merchant/assets`
-  const timestamp = String(Date.now())
-  const signature = await generateSignature(
-    process.env.FONBNK_CLIENT_SECRET,
-    timestamp,
-    '/api/pay-widget-merchant/assets',
-  )
-  const requestOptions = {
-    method: 'GET',
-    headers: {
-      'x-client-id': FONBNK_CLIENT_ID,
-      'x-timestamp': timestamp,
-      'x-signature': signature,
-    },
-  }
-
-  const response = await fetchWithBackoff(url, requestOptions)
-
-  if (!response.ok) {
-    if (response.status === 404) {
-      return []
-    }
-    throw new Error(`Error fetching fonbnk assets: ${response.statusText}`)
-  }
-
-  const data: FonbnkAsset[] = await response.json()
-  return data
-}
-
-async function getPayoutWallets({
-  fonbnkNetwork,
-  currency,
-}: {
-  fonbnkNetwork: FonbnkNetwork
-  currency: string
-}): Promise<string[]> {
-  const url = `${FONBNK_API_URL}/api/util/payout-wallets?network=${fonbnkNetwork}&asset=${currency}`
-  const timestamp = String(Date.now())
-  const signature = await generateSignature(
-    process.env.FONBNK_CLIENT_SECRET,
-    timestamp,
-    '/api/util/payout-wallets?network=${fonbnkNetwork}&asset=${currency}',
-  )
-  const requestOptions = {
-    method: 'GET',
-    headers: {
-      'x-client-id': FONBNK_CLIENT_ID,
-      'x-timestamp': timestamp,
-      'x-signature': signature,
-    },
-  }
-
-  const response = await fetchWithBackoff(url, requestOptions)
-
-  if (!response.ok) {
-    if (response.status === 404) {
-      return []
-    }
-    throw new Error(`Error fetching fonbnk assets: ${response.statusText}`)
-  }
-  const data: string[] = await response.json()
-  return data
-}
+import { AERODROME_NETWORK_ID } from '../aerodrome/constants'
+import { FonbnkTransaction } from './types'
 
 async function getUserTransactions({
   address,
@@ -131,7 +53,7 @@ async function getTotalRevenueUsdFromTransactions({
   transactions,
   networkId,
   startTimestamp,
-  endTimestamp
+  endTimestamp,
 }: {
   transactions: FonbnkTransaction[]
   networkId: NetworkId
@@ -143,7 +65,10 @@ async function getTotalRevenueUsdFromTransactions({
   }
   let totalUsdContribution = 0
   const tokenId = `${networkId}:${transactions[0].tokenAddress}`
-  const tokenContract = await getErc20Contract(transactions[0].tokenAddress, networkId)
+  const tokenContract = await getErc20Contract(
+    transactions[0].tokenAddress,
+    networkId,
+  )
   const tokenDecimals = BigInt(await tokenContract.read.decimals())
   const tokenPrices = await fetchTokenPrices({
     tokenId,
