+++ scripts/calculateRevenue/protocols/fonbnk/types.ts
@@ -1,0 +1,13 @@
+import { Address } from "viem"
+import { FonbnkNetwork } from "./constants"
+
+export interface FonbnkAsset {
+    network: FonbnkNetwork
+    asset: string
+  }
+  
+export interface FonbnkTransaction {
+    amount: bigint
+    tokenAddress: Address
+    timestamp: Date
+  }
